# perf(profiling): Tune the sample profile generation code for performance