@Zylphrex 2022-10-19
Some benchmarks I collected comparing the before and afters using a few arbitrary profiles.

```samples: 439		stacks: 4326
old	avg of 100 runs	   21.9256ms
new	avg of 100 runs	    3.5280ms
samples: 507		stacks: 4227
old	avg of 100 runs	   18.8779ms
new	avg of 100 runs	    4.2562ms
samples: 1688		stacks: 15167
old	avg of 100 runs	   61.5474ms
new	avg of 100 runs	   14.6774ms
samples: 418		stacks: 3347
old	avg of 100 runs	   14.5421ms
new	avg of 100 runs	    2.5589ms
samples: 513		stacks: 2568
old	avg of 100 runs	   14.5687ms
new	avg of 100 runs	    2.9569ms
samples: 379		stacks: 758
old	avg of 100 runs	    5.7061ms
new	avg of 100 runs	    1.5271ms
```

From this testing, this looks to be consistently faster than what it used to be, and I have no reason to believe it's slower in any case.