@antonpirker 2022-10-20
Side note: The "daphne" errors above in the Django tests was fixed in master, to if you update your branches they should be gone.